requirements.txt
pandas==1.0.5
dash==1.13.4
plotly==4.4.1
mysql-connector-python==8.0.21
numpy~=1.18.3
sklearn~=0.0
scikit-learn~=0.21.3
umap-learn==0.4.5
python-dotenv~=0.12.0